Adjusted URL parsing & Single clear reports

+++ bot.py
@@ -35,15 +35,19 @@ async def on_ready():
 async def rankings(ctx, log_link):
     # Parse the log link
     parsed_url = urlparse(log_link)
-
+    parse_path = parsed_url.path.split('/')
+    
+    print(parse_path)
+    print(parse_path[2])
+    
     # Assert the first path segment is "reports"
-    if parsed_url.path.split('/')[1] != 'reports':
+    if parse_path[1] != 'reports':
         await ctx.send('Invalid log link / Not a report link')
         return
 
     # Get the last location of the path
-    report_id = parsed_url.path.split('/')[-1]
-
+    report_id = parse_path[2]
+    
     # Query
     q = '''
     {
@@ -62,51 +66,56 @@ async def rankings(ctx, log_link):
 
     res = ff_client.query(q)
 
-    # Check < 1 clears
-    if not res['data']['reportData']['report']['rankings']['data']:
-        await ctx.send('`No rankings found`')
-        return
+    # if 'errors' in res: 
 
     # Build time dict
     time_dict = {f['id'] : f['endTime'] - f['startTime'] for f in res['data']['reportData']['report']['fights']}
     # Build rankings dict
     rankings_dict = {f['fightID'] : f for f in res['data']['reportData']['report']['rankings']['data']}
 
-    # Prompt user to select a fight
-    tmp_table = []
-    for fight in res['data']['reportData']['report']['rankings']['data']:
-        tmp_table.append([fight['fightID'], fight['encounter']['name'], ms_to_mmssms(time_dict[fight['fightID']])])
-    
-    pmt_msg = await ctx.send("`Enter the fight number you want to see rankings for\n"
-        + tabulate(tmp_table, headers=['Fight Number', 'Encounter', 'Time'], tablefmt='github')
-        + "`"
-    )
-
-    # Validate user input
-    def response_check(msg):
-        return msg.author == ctx.author and msg.channel == ctx.channel
-
-    try:
-        msg = await bot.wait_for('message', check=response_check, timeout=30.0)
-        i = int(msg.content)
+    # Check < 1 clears
+    clear_count = len(res['data']['reportData']['report']['rankings']['data'])
+    if clear_count == 0:
+        await ctx.send('`No rankings found`')
+        return
+    elif clear_count == 1:
+        fight = list(rankings_dict.values())[0]
+        i = fight['fightID']
+    else:
+        # Prompt user to select a fight
+        tmp_table = []
+        for fight in res['data']['reportData']['report']['rankings']['data']:
+            tmp_table.append([fight['fightID'], fight['encounter']['name'], ms_to_mmssms(time_dict[fight['fightID']])])
         
-        # Select the fight
-        fight = rankings_dict[i]
-        
-    except (IndexError, KeyError):
-        await ctx.send("`Invalid fight number`")
+        pmt_msg = await ctx.send("`Enter the fight number you want to see rankings for\n"
+            + tabulate(tmp_table, headers=['Fight Number', 'Encounter', 'Time'], tablefmt='github')
+            + "`"
+        )
+
+        # Validate user input
+        def response_check(msg):
+            return msg.author == ctx.author and msg.channel == ctx.channel
+
+        try:
+            msg = await bot.wait_for('message', check=response_check, timeout=30.0)
+            i = int(msg.content)
+            
+            # Select the fight
+            fight = rankings_dict[i]
+        except (IndexError, KeyError):
+            await ctx.send("`Invalid fight number`")
+            await msg.delete()
+            return
+        except asyncio.TimeoutError:
+            await ctx.send('`Request timed out`')
+            return
+        except ValueError:
+            await ctx.send('`Invalid input`')
+            await msg.delete()
+            return
+        finally:
+            await pmt_msg.delete()
         await msg.delete()
-        return
-    except asyncio.TimeoutError:
-        await ctx.send('`Request timed out`')
-        return
-    except ValueError:
-        await ctx.send('`Invalid input`')
-        await msg.delete()
-        return
-    finally:
-        await pmt_msg.delete()
-    await msg.delete()
     
     
     # Format and send
